[Task] Add next function & previous function & slide function v3.4
