+++ App_Music/Constants.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+enum Constants: String{
+    case lastPlaybackTime = "LastPlaybackTime"
+    case lastSongIndex    = "LastSongIndex"
+    case mp3              = "mp3"
+    case jpg              = "jpg"
+}
