+++ App_Music/Song.swift
@@ -1,13 +1,6 @@
-//
-//  Song.swift
-//  App_Music
-//
-//  Created by Thai Nguyen Khac on 03/07/2023.
-//
-
 import UIKit
 
-class Song {
+final class Song {
     var songName:String
     var songImage:UIImage
     var songArtist:String
